src/data-injection/requirements.txt
aws-cdk-lib==2.94.0
constructs>=10.0.0,<11.0.0
boto3
requests==2.26.0
pandas
sqlalchemy